55: extract `<About>` and its test into separate files
