+++ frontend/src/features/About.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+export const About = () => {
+  console.log(`${new Date().toISOString()} - React is rendering <About>`);
+
+  return (
+    <React.Fragment>
+      {"<About>"}
+      <h1>About VocabTreasury...</h1>
+    </React.Fragment>
+  );
+};
